+++ src/Domain/Repository/ChatSessionRepositoryInterface.php
@@ -4,6 +4,7 @@ namespace Proximum\Vimeet\Domain\Repository;
 
 use Proximum\Vimeet\Domain\Model\ChatSession;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
 
 interface ChatSessionRepositoryInterface
@@ -24,4 +25,6 @@ interface ChatSessionRepositoryInterface
     public function findOneById(int $id): ?ChatSession;
 
     public function hasMessageFromUser (ChatSession $chatSession, User $user): bool;
+
+    public function hasAStartedVisio (Event $event, User $user): bool;
 }